make run go test serially
